pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "AI Desk Chat Application Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "python-dotenv==1.0.0",
    "aiofiles==23.2.1",
    "sqlmodel==0.0.14",
    "sqlalchemy==2.0.23",
    "google-generativeai==0.8.3",
    "python-jose[cryptography]==3.3.0",
    "argon2-cffi==23.1.0",
    "python-multipart==0.0.20",
]

[tool.pytest.ini_options]
addopts = "-v --tb=short --html=tests/report.html --self-contained-html"
testpaths = ["tests"]
markers = ["integration: marks tests as integration tests"]

[tool.uv]
dev-dependencies = [
    "pytest==7.4.3",
    "pytest-asyncio==0.23.2",
    "pytest-html==4.1.1",
    "httpx==0.25.2",
]
